Refactor db initializer path
